requirements.txt
python-telegram==0.19.0
psycopg2-binary~=2.9.10
python-dotenv~=1.1.1
loguru~=0.7.3